refactor: lcd, rpc uses same logic
