+++ src/lib/rpc.ts
@@ -1,31 +1,11 @@
-import { request, Agent } from 'undici'
 import { compact } from 'lodash'
 import config from 'config'
+import { request } from 'lib/request'
 import { apiLogger as logger } from './logger'
 
-const agent = new Agent({
-  connect: {
-    rejectUnauthorized: false
-  }
-})
-
 async function getRequest(path: string, params?: Record<string, string>): Promise<any> {
-  const options = {
-    headers: {
-      'Content-Type': 'application/json',
-      'User-Agent': 'terra-fcd'
-    },
-    dispatcher: agent
-  }
-
-  let url = `${config.RPC_URI}${path}`
-  params && Object.keys(params).forEach((key) => params[key] === undefined && delete params[key])
-  const qs = new URLSearchParams(params as any).toString()
-  if (qs.length) {
-    url += `?${qs}`
-  }
-
-  const response = await request(url, options)
+  const url = `${config.RPC_URI}${path}`
+  const response = await request(url, params)
     .then((res) => {
       if (res.statusCode !== 200) {
         throw new Error('invalid status')
